% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bridge}
\alias{bridge}
\title{Bayesian Ridge prior}
\usage{
bridge()
}
\value{
A prior specification object of class \code{c("prior_bridge", "bpm_prior")}.
}
\description{
Estimates a global ridge (L2) penalty on all slope coefficients via REML,
using \code{mgcv::gam} with \code{paraPen}. Predictors are standardised internally;
coefficients are returned on the original (unstandardised) scale.
}
