% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{coef.bpm_proj_pm}
\alias{coef.bpm_proj_pm}
\title{Extract coefficients from a bpm_proj_pm object}
\usage{
\method{coef}{bpm_proj_pm}(object, ...)
}
\arguments{
\item{object}{A \code{bpm_proj_pm} object.}

\item{...}{Ignored.}
}
\description{
Extract coefficients from a bpm_proj_pm object
}
