% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.bpmfit}
\alias{coef.bpmfit}
\title{Extract coefficients from a bpm object}
\usage{
\method{coef}{bpmfit}(object, ...)
}
\arguments{
\item{object}{A \code{bpmfit} object.}

\item{...}{Ignored.}
}
\description{
Extract coefficients from a bpm object
}
