% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{coef.bpmproj_pm}
\alias{coef.bpmproj_pm}
\title{Extract coefficients from a bpmproj_pm object}
\usage{
\method{coef}{bpmproj_pm}(object, ...)
}
\arguments{
\item{object}{A \code{bpmproj_pm} object.}

\item{...}{Ignored.}
}
\description{
Extract coefficients from a bpmproj_pm object
}
