% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{flat}
\alias{flat}
\title{Flat (improper) prior}
\usage{
flat()
}
\value{
A prior specification object of class \code{c("prior_flat", "bpm_prior")}.
}
\description{
Equivalent to unpenalised logistic regression via \code{glm()}.
}
