% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{jeffreys}
\alias{jeffreys}
\title{Jeffreys prior}
\usage{
jeffreys()
}
\value{
A prior specification object of class \code{c("prior_jeffreys", "bpm_prior")}.
}
\description{
Fitted via \code{brglm2::brglm_fit} with \code{type = "AS_mean"}, which implements the
mean bias-reducing adjusted scores (Firth 1993; Kosmidis & Firth 2021).
This is equivalent to a Jeffreys prior on the coefficients.
}
