% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{predict.bpmproj_pm}
\alias{predict.bpmproj_pm}
\title{Predict from a PM-projected model}
\usage{
\method{predict}{bpmproj_pm}(
  object,
  newdata,
  type = c("response", "link"),
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{A \code{bpmproj_pm} object returned by \code{\link[=bpmproj_pm]{bpmproj_pm()}}.}

\item{newdata}{A data frame of new observations.}

\item{type}{\code{"response"} (default) for predicted probability;
\code{"link"} for the linear predictor.}

\item{na.action}{Function to handle \code{NA}s in \code{newdata}. Default
\link[stats:na.pass]{stats::na.pass}.}

\item{...}{Currently unused.}
}
\value{
A numeric vector.
}
\description{
Predict from a PM-projected model
}
\seealso{
\code{\link[=bpmproj_pm]{bpmproj_pm()}}
}
