% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{print.bpm_proj_pm}
\alias{print.bpm_proj_pm}
\title{Print a bpm_proj_pm object}
\usage{
\method{print}{bpm_proj_pm}(x, ...)
}
\arguments{
\item{x}{A \code{bpm_proj_pm} object.}

\item{...}{Ignored.}
}
\description{
Print a bpm_proj_pm object
}
