% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bpmfit}
\alias{print.bpmfit}
\title{Print a bpmfit object}
\usage{
\method{print}{bpmfit}(x, ...)
}
\arguments{
\item{x}{A \code{bpmfit} object.}

\item{...}{Ignored.}
}
\description{
Print a bpmfit object
}
