% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{print.bpmproj_pm}
\alias{print.bpmproj_pm}
\title{Print a bpmproj_pm object}
\usage{
\method{print}{bpmproj_pm}(x, ...)
}
\arguments{
\item{x}{A \code{bpmproj_pm} object.}

\item{...}{Ignored.}
}
\description{
Print a bpmproj_pm object
}
