% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.bpm}
\alias{vcov.bpm}
\title{Extract the posterior covariance matrix from a bpm object}
\usage{
\method{vcov}{bpm}(object, ...)
}
\arguments{
\item{object}{A \code{bpm} object.}

\item{...}{Ignored.}
}
\description{
Extract the posterior covariance matrix from a bpm object
}
