% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.bpmfit}
\alias{vcov.bpmfit}
\title{Extract the posterior covariance matrix from a bpmfit object}
\usage{
\method{vcov}{bpmfit}(object, ...)
}
\arguments{
\item{object}{A \code{bpmfit} object.}

\item{...}{Ignored.}
}
\description{
Extract the posterior covariance matrix from a bpmfit object
}
